aitesms/requirements.txt
PySide6==6.10.1
PyYAML==6.0.2
